backend/requirements.txt
fastapi
uvicorn
httpx
pydantic
python-multipart
requests
google-genai